put the pee_id from network module to network-p2p for decoupling
