+++ cmd/starcoin/src/node/network/known_peers_cmd.rs
@@ -5,7 +5,7 @@ use crate::cli_state::CliState;
 use crate::StarcoinOpt;
 use anyhow::Result;
 use clap::Parser;
-use network_types::peer_info::PeerId;
+use network_p2p_types::peer_id::PeerId;
 use scmd::{CommandAction, ExecContext};
 
 #[derive(Debug, Parser, Default)]
