+++ network-p2p/core/Cargo.toml
@@ -7,7 +7,7 @@ network-p2p-derive = { workspace = true }
 num_enum = { workspace = true }
 serde = { features = ["derive"], workspace = true }
 starcoin-types = { package = "starcoin-types", workspace = true }
-network-types = { workspace = true }
+network-p2p-types = { workspace = true }
 
 [dev-dependencies]
 stest = { workspace = true }
