+++ network-p2p/types/Cargo.toml
@@ -11,6 +11,8 @@ schemars = { workspace = true }
 serde = { features = ["derive"], workspace = true }
 serde_json = { features = ["arbitrary_precision"], workspace = true }
 anyhow = { workspace = true }
+starcoin-crypto = { workspace = true }
+starcoin-types = { workspace = true }
 
 [features]
 default = []
