+++ network-rpc/api/Cargo.toml
@@ -5,6 +5,7 @@ bcs-ext = { package = "bcs-ext", workspace = true }
 futures = { workspace = true }
 network-p2p-core = { workspace = true }
 network-p2p-derive = { workspace = true }
+network-p2p-types = { workspace = true }
 network-types = { workspace = true }
 serde = { features = ["derive"], workspace = true }
 starcoin-accumulator = { workspace = true }
