+++ network-rpc/api/src/lib.rs
@@ -5,7 +5,8 @@ use anyhow::Result;
 use futures::future::BoxFuture;
 use network_p2p_core::{NetRpcError, RpcErrorCode};
 use network_p2p_derive::*;
-use network_types::peer_info::{PeerId, RpcInfo};
+use network_p2p_types::peer_id::PeerId;
+use network_types::peer_info::RpcInfo;
 use once_cell::sync::Lazy;
 use serde::{Deserialize, Serialize};
 use starcoin_accumulator::node::AccumulatorStoreType;
