+++ network/api/src/messages.rs
@@ -5,7 +5,8 @@ use crate::ReputationChange;
 use anyhow::*;
 use bcs_ext::{BCSCodec, Sample};
 use futures::channel::oneshot::Receiver;
-use network_types::peer_info::{PeerId, PeerInfo};
+use network_p2p_types::peer_id::PeerId;
+use network_types::peer_info::PeerInfo;
 use serde::{Deserialize, Serialize};
 use starcoin_crypto::HashValue;
 use starcoin_service_registry::ServiceRequest;
