+++ network/src/lib.rs
@@ -13,6 +13,7 @@ pub mod worker;
 pub use network_api::messages::*;
 
 pub use helper::{get_unix_ts, get_unix_ts_as_millis};
+pub use network_p2p_types::peer_id::PeerId;
 pub use service::NetworkActorService;
 pub use service_ref::NetworkServiceRef;
 pub use worker::build_network_worker;
