+++ rpc/api/src/types.rs
@@ -6,7 +6,8 @@ use bcs_ext::BCSCodec;
 use hex::FromHex;
 use jsonrpc_core_client::RpcChannel;
 use move_core_types::u256;
-use network_types::peer_info::{PeerId, PeerInfo};
+use network_p2p_types::peer_id::PeerId;
+use network_types::peer_info::PeerInfo;
 pub use node_api_types::*;
 use schemars::{self, JsonSchema};
 use serde::de::{DeserializeOwned, Error};
