+++ rpc/client/src/lib.rs
@@ -14,7 +14,8 @@ pub use jsonrpc_core::Params;
 use jsonrpc_core_client::{transports::ipc, transports::ws, RpcChannel};
 use network_api::PeerStrategy;
 use network_p2p_types::network_state::NetworkState;
-use network_types::peer_info::{Multiaddr, PeerId};
+use network_p2p_types::peer_id::PeerId;
+use network_types::peer_info::Multiaddr;
 use parking_lot::Mutex;
 use serde_json::Value;
 use starcoin_abi_types::{FunctionABI, ModuleABI, StructInstantiation};
