+++ rpc/server/src/module/network_manager_rpc.rs
@@ -8,7 +8,8 @@ use jsonrpc_core::Result;
 use network_api::{PeerProvider, ReputationChange, BANNED_THRESHOLD};
 use network_p2p_core::RawRpcClient;
 use network_p2p_types::network_state::NetworkState;
-use network_types::peer_info::{Multiaddr, PeerId};
+use network_p2p_types::peer_id::PeerId;
+use network_types::peer_info::Multiaddr;
 use starcoin_network::NetworkServiceRef;
 use starcoin_rpc_api::network_manager::NetworkManagerApi;
 use starcoin_rpc_api::types::StrView;
